Lazily load the model, construct the client once in the constructor
